Moved plug example to the plug folder and update plug disconnect and power off structure
