+++ openhtf/plugs/acroname/acroname_power_supply_example.yaml
